moved save project to SaveProjectScenario
