+++ src/project/projectmodule.cpp
@@ -29,6 +29,7 @@
 #include "internal/projectuiactions.h"
 #include "internal/projectconfiguration.h"
 #include "internal/opensaveprojectscenario.h"
+#include "internal/saveprojectscenario.h"
 #include "internal/exportprojectscenario.h"
 #include "internal/mscmetareader.h"
 #include "internal/templatesrepository.h"
@@ -129,6 +130,7 @@ void ProjectContext::registerExports()
     ioc()->registerExport<IProjectFilesController>(mname, m_actionsController);
     ioc()->registerExport<mi::IProjectProvider>(mname, m_actionsController);
     ioc()->registerExport<IOpenSaveProjectScenario>(mname, new OpenSaveProjectScenario(iocContext()));
+    ioc()->registerExport<ISaveProjectScenario>(mname, new SaveProjectScenario(iocContext()));
     ioc()->registerExport<IExportProjectScenario>(mname, new ExportProjectScenario(iocContext()));
     ioc()->registerExport<IRecentFilesController>(mname, m_recentFilesController);
     ioc()->registerExport<ITemplatesRepository>(mname, new TemplatesRepository());
