+++ src/project/tests/CMakeLists.txt
@@ -28,7 +28,7 @@ set(MODULE_TEST_SRC
     ${CMAKE_CURRENT_LIST_DIR}/mocks/projectconfigurationmock.h
     ${CMAKE_CURRENT_LIST_DIR}/mocks/recentfilescontrollermock.h
     ${CMAKE_CURRENT_LIST_DIR}/mscmetareadertests.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/projectactionscontrollertests.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/saveprojectscenariotests.cpp
     ${CMAKE_CURRENT_LIST_DIR}/templatesrepositorytest.cpp
 )
